requirements.txt
openai
python-dotenv
langchain
tqdm